requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0
dashscope>=1.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
